--- src/utils/country.tsx
@@ -1,19 +1,0 @@
-import countriesJSON from '../assets/countries.json';
-
-const countries: {
-    [key: string]: {
-        name: string;
-        native: string;
-        currency: string;
-        languages: string[];
-        emoji: string;
-    };
-} = countriesJSON;
-
-export function getCountryNameFromInitials(countryInitials: string) {
-    if (countries[countryInitials]) {
-        return countries[countryInitials].name;
-    }
-
-    return 'Unknown';
-}
