+++ src/views/Stories/Filters.tsx
@@ -1,21 +1,20 @@
 import { Col, Row } from '@impact-market/ui';
 import { countriesOptions } from '../../utils/countries';
-import FilterSelect from '../../components/FilterSelect';
 import React from 'react';
+import Select from '../../components/Select';
 import useFilters from '../../hooks/useFilters';
 
 const Filters = () => {
-    // const { clear, update } = useFilters();
     const { update } = useFilters();
 
     const updateFilter  = (value: any) => {
-            update('country', [value]);  
+        update('country', [value]);  
     }
 
     return (
         <Row mt={0.625}>
             <Col colSize={{ sm: 6, xs: 12}}>
-                <FilterSelect
+                <Select
                     callback={updateFilter}
                     isMultiple
                     options={countriesOptions}
